refactor: add framework hexagon
